+++ backend/libraries/types/src/events.rs
@@ -34,6 +34,7 @@ pub enum ChatEvent {
     GroupGateUpdated(GroupGateUpdated),
     UsersInvited(UsersInvited),
     MembersAddedToDefaultChannel(MembersAddedToDefaultChannel),
+    ExternalUrlUpdated(ExternalUrlUpdated),
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug, TS)]
@@ -254,6 +255,12 @@ pub struct EventsTimeToLiveUpdated {
 pub struct GroupGateUpdated {
     pub updated_by: UserId,
     pub new_gate: Option<AccessGate>,
+}
+
+#[derive(CandidType, Serialize, Deserialize, Clone, Debug, TS)]
+pub struct ExternalUrlUpdated {
+    pub updated_by: UserId,
+    pub new_url: Option<String>,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Copy, Clone, Debug, TS)]