+++ frontend/openchat-worker/src/worker.ts
@@ -496,7 +496,7 @@ self.addEventListener("message", (msg: MessageEvent<CorrelatedWorkerRequest>) =>
                 executeThenReply(
                     payload,
                     correlationId,
-                    agent.joinGroup(payload.chatId, payload.localUserIndex, payload.credentialArgs),
+                    agent.joinGroup(payload.chatId, payload.credentialArgs),
                 );
                 break;
 
@@ -504,7 +504,7 @@ self.addEventListener("message", (msg: MessageEvent<CorrelatedWorkerRequest>) =>
                 executeThenReply(
                     payload,
                     correlationId,
-                    agent.joinCommunity(payload.id, payload.localUserIndex, payload.credentialArgs),
+                    agent.joinCommunity(payload.id, payload.credentialArgs),
                 );
                 break;
 
@@ -715,12 +715,7 @@ self.addEventListener("message", (msg: MessageEvent<CorrelatedWorkerRequest>) =>
                 executeThenReply(
                     payload,
                     correlationId,
-                    agent.inviteUsers(
-                        payload.chatId,
-                        payload.localUserIndex,
-                        payload.userIds,
-                        payload.callerUsername,
-                    ),
+                    agent.inviteUsers(payload.chatId, payload.userIds, payload.callerUsername),
                 );
                 break;
 
@@ -730,7 +725,6 @@ self.addEventListener("message", (msg: MessageEvent<CorrelatedWorkerRequest>) =>
                     correlationId,
                     agent.inviteUsersToCommunity(
                         payload.id,
-                        payload.localUserIndex,
                         payload.userIds,
                         payload.callerUsername,
                     ),